+++ plugin-build/src/test/kotlin/androidx/paging/PagingSource.kt
@@ -1,0 +1,3 @@
+package androidx.paging
+
+abstract class PagingSource<Key : Any, Value : Any>