+++ plugin-build/src/test/kotlin/androidx/room/RoomSQLiteQuery.kt
@@ -1,0 +1,6 @@
+package androidx.room
+
+import androidx.sqlite.db.SupportSQLiteProgram
+import androidx.sqlite.db.SupportSQLiteQuery
+
+abstract class RoomSQLiteQuery : SupportSQLiteQuery, SupportSQLiteProgram