+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginVariantTest.kt
@@ -188,6 +188,9 @@ class SentryPluginVariantTest(
                   ignoredVariants = [$variants]
                   ignoredBuildTypes = [$buildTypes]
                   ignoredFlavors = [$flavors]
+                  tracingInstrumentation {
+                    enabled = false
+                  }
                 }
             """.trimIndent()
         )
@@ -201,15 +204,11 @@ class SentryPluginVariantTest(
             // The supported Gradle version can be found here:
             // https://developer.android.com/studio/releases/gradle-plugin#updating-gradle
             // The pair is [AGP Version, Gradle Version]
-            arrayOf("4.0.2", "6.1.1"),
-            arrayOf("4.1.3", "6.5"),
-            arrayOf("4.1.3", "6.8.3"),
-            arrayOf("4.1.3", "7.0.2"),
-            arrayOf("4.2.2", "6.8.3"),
-            arrayOf("4.2.2", "7.0.2"),
-            arrayOf("7.0.2", "7.0.2"),
-            arrayOf("7.0.2", "7.1.1"),
-            arrayOf("7.0.2", "7.2")
+            arrayOf("7.0.3", "7.0.2"),
+            arrayOf("7.0.3", "7.1.1"),
+            arrayOf("7.0.3", "7.2"),
+            arrayOf("7.1.0-beta01", "7.2"),
+            arrayOf("7.2.0-alpha01", "7.2")
         )
 
         private fun GradleRunner.appendArguments(vararg arguments: String) =